ml-service/requirements.txt
# ML Service Dependencies for POS Myanmar AI Features
# Python 3.10+

# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Time Series Forecasting
prophet==1.1.4

# Association Rules / Market Basket Analysis
mlxtend==0.22.0

# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Database
mysql-connector-python==8.1.0

# Model Persistence
joblib==1.3.2

# Data Visualization (for notebooks)
matplotlib==3.7.2
seaborn==0.12.2

# Jupyter for EDA
jupyter==1.0.0
notebook==7.0.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
